📊Stats! (#14)
This PR adds stats to kube-service-exporter in a couple different forms:

* Via sending metrics to Dogstatsd
  * consul RPC success/failure/timings
  * frequency of kubernetes watches firing
* Internal go memstats via go_expvar, exposed to datadog via DD service discovery

In order to make the important logic more readable, I opted to *wrap* the Consul API calls w/ `InstrumentedKV` and `InstrumentedAgent`.  This makes things more legible, and improves the odds that a call will fail to be instrumented.

Longer term, I'd like to switch much of this over to prometheus metrics (primarily for open source purposes), but (dog)statsd was lower friction at this stage.

Because I had to add an http listener already for go_expvar, I also added a readiness probe at `/healthz` which does nothing but return ok.  It should probably be made smarter, but this PR is already pretty big.