use container port when port is unnamed
In cases where the container port is unnamed, KSE currently uses the
NodePort as a placeholder in the canonical service ID. However NodePort
is likely to be unique per cluster, so this breaks any assumptions
around multi-cluster load balancing.

This PR instead uses the port number (which is a better analog for the
port name) for the canonical service Id, which will fix multi-cluster
load balancing in cases where the port name is unspecified.